+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Version [6.1.2][unreleased] - (in development)
+- New: Add Integration Tests
 
 ## Version [6.1.1] - (2016-01-11)
 - Fixed: Wrong name of syncurl in tests